Update vers V2.0 : organisation du code dans plusieurs fichiers, optimisation générale, ajout de commentaires.
