--- manifest.json
@@ -1,20 +1,0 @@
-{
-    "manifest_version": 3,
-    "name": "Arckchually Corrector",
-    "short_name": "Arckchually",
-    "version": "2.0",
-    "description": "Corrige les fautes de français et affiche un message '☝️🤓 Erm, arckchually...' pour aider à l'orthographe.",
-  
-    "host_permissions": [
-      "https://raw.githubusercontent.com/*"
-    ],
-
-    "content_scripts": [
-      {
-        "matches": ["<all_urls>"],
-        "js": ["content.js"],
-        "run_at": "document_end"
-      }
-    ]
-  }
-
