+++ tsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "target": "es6", // Version cible de JavaScript
+        "module": "commonjs", // Système de module
+        "strict": true, // Activer les vérifications strictes
+        "esModuleInterop": true, // Interopérabilité avec les modules ES
+        "skipLibCheck": true, // Ignorer la vérification des fichiers de définition
+        "forceConsistentCasingInFileNames": true, // Forcer la casse cohérente dans les noms de fichiers
+        "outDir": "./dist", // Répertoire de sortie
+        "rootDir": "./src" // Répertoire racine
+    },
+    "include": [
+        "src/**/*" // Inclure tous les fichiers dans src
+    ],
+    "exclude": [
+        "node_modules", // Exclure le répertoire node_modules
+        "**/*.spec.ts" // Exclure les fichiers de test
+    ]
+}
